requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
aiofiles==24.1.0
asyncpg==0.29.0
pgvector==0.3.5
sentence-transformers==3.3.0
torch==2.5.1
numpy==1.26.4
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-community==0.3.7
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
markdown==3.7
beautifulsoup4==4.12.3
chardet==5.2.0
aiohttp==3.11.7

